requirements.txt
pandas
selenium==4.15.2
webdriver-manager==4.0.1
requests==2.31.0
beautifulsoup4==4.12.2
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
aiohttp==3.9.1
statsbombpy==1.11.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-timeout==2.2.0
click==8.1.7      # o la última versión
